Select native targets by platform
