+++ CMakeLists.txt
@@ -2,20 +2,10 @@ cmake_minimum_required(VERSION 3.18)
 
 project(LupineProject LANGUAGES C CXX)
 
-set(LUPINE_BUILD_SERVER_DEFAULT ON)
-set(LUPINE_BUILD_NVML_CLIENT_DEFAULT ON)
-if(APPLE)
-  # NVIDIA does not ship a Darwin CUDA toolkit or driver, but the RPC client
-  # only needs the platform-independent CUDA headers.
-  set(LUPINE_BUILD_SERVER_DEFAULT OFF)
-  set(LUPINE_BUILD_NVML_CLIENT_DEFAULT OFF)
-endif()
-option(LUPINE_BUILD_SERVER "Build the CUDA driver server"
-       ${LUPINE_BUILD_SERVER_DEFAULT})
-option(LUPINE_BUILD_NVML_CLIENT "Build the NVML client shim"
-       ${LUPINE_BUILD_NVML_CLIENT_DEFAULT})
 set(LUPINE_CUDA_INCLUDE_DIR "" CACHE PATH
-    "CUDA include directory for a header-only client build")
+    "CUDA include directory for a macOS client build")
+set(LUPINE_NVML_INCLUDE_DIR "" CACHE PATH
+    "NVML include directory for a macOS client build")
 
 # Sanitizers are applied only to the CPU-only *_test executables (transport,
 # decoder, ipc, checkpoint), never the nvcc-compiled driver/server. Set to a
@@ -43,28 +33,34 @@ function(lupine_scaled_timeout out base)
     set(${out} ${scaled} PARENT_SCOPE)
 endfunction()
 
-if (LUPINE_BUILD_SERVER AND NOT WIN32)
+if (NOT WIN32 AND NOT APPLE)
     enable_language(CUDA)
 endif()
 
 set(MIN_CUDA_VERSION 11.0)
 
-if(LUPINE_CUDA_INCLUDE_DIR)
+if(APPLE)
+  if(NOT LUPINE_CUDA_INCLUDE_DIR OR NOT LUPINE_NVML_INCLUDE_DIR)
+    message(FATAL_ERROR
+        "macOS client builds require LUPINE_CUDA_INCLUDE_DIR and LUPINE_NVML_INCLUDE_DIR")
+  endif()
   find_path(LUPINE_CUDA_HEADER_DIR cuda.h
       PATHS "${LUPINE_CUDA_INCLUDE_DIR}"
       NO_DEFAULT_PATH
       REQUIRED)
-  set(CUDAToolkit_INCLUDE_DIRS "${LUPINE_CUDA_HEADER_DIR}")
-  if(LUPINE_BUILD_SERVER)
-    message(FATAL_ERROR
-        "LUPINE_CUDA_INCLUDE_DIR supports client-only builds; the server requires a full CUDA toolkit")
-  endif()
+  find_path(LUPINE_NVML_HEADER_DIR nvml.h
+      PATHS "${LUPINE_NVML_INCLUDE_DIR}"
+      NO_DEFAULT_PATH
+      REQUIRED)
+  set(CUDAToolkit_INCLUDE_DIRS
+      "${LUPINE_CUDA_HEADER_DIR}"
+      "${LUPINE_NVML_HEADER_DIR}")
   file(STRINGS "${LUPINE_CUDA_HEADER_DIR}/cuda.h" LUPINE_CUDA_VERSION_LINE
        REGEX "^#define CUDA_VERSION +[0-9]+$")
   string(REGEX REPLACE ".*CUDA_VERSION +([0-9]+).*" "\\1"
          LUPINE_CUDA_HEADER_VERSION "${LUPINE_CUDA_VERSION_LINE}")
   message(STATUS
-      "Using header-only CUDA API version: ${LUPINE_CUDA_HEADER_VERSION}")
+      "Using macOS CUDA API headers version: ${LUPINE_CUDA_HEADER_VERSION}")
 else()
   find_package(CUDAToolkit REQUIRED)
 
@@ -86,7 +82,7 @@ find_package(OpenSSL QUIET)
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
-if (LUPINE_BUILD_SERVER AND NOT WIN32)
+if (NOT WIN32 AND NOT APPLE)
     set(CMAKE_CUDA_STANDARD 14)
     set(CMAKE_CUDA_STANDARD_REQUIRED ON)
     set(CMAKE_CUDA_EXTENSIONS OFF)
@@ -152,7 +148,6 @@ set(CUDA_CLIENT_HEADERS
     ${CMAKE_CURRENT_SOURCE_DIR}/checkpoint.h
     ${CMAKE_CURRENT_SOURCE_DIR}/client_routing.h
     ${CMAKE_CURRENT_SOURCE_DIR}/ipc.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/lupine_elf.h
     ${CMAKE_CURRENT_SOURCE_DIR}/codegen/gen_cuda_client.h
     ${CMAKE_CURRENT_SOURCE_DIR}/memcpy.h
     ${CMAKE_CURRENT_SOURCE_DIR}/copy_pipeline.h
@@ -204,30 +199,32 @@ if (NOT WIN32)
       )
     endif()
 
-    if(LUPINE_BUILD_NVML_CLIENT)
-      add_library(${NVML_CLIENT_OUTPUT} SHARED ${NVML_CLIENT_SOURCES})
-      target_include_directories(${NVML_CLIENT_OUTPUT} PRIVATE ${CUDAToolkit_INCLUDE_DIRS})
-      if(OpenSSL_FOUND)
-        target_compile_definitions(${NVML_CLIENT_OUTPUT} PRIVATE LUPINE_TLS_OPENSSL)
-        target_link_libraries(${NVML_CLIENT_OUTPUT} PRIVATE OpenSSL::SSL OpenSSL::Crypto)
-      endif()
-      target_link_libraries(${NVML_CLIENT_OUTPUT} PRIVATE ${NGHTTP2_LIBRARY} lupine_lz4)
-      if(NOT APPLE)
-        target_link_libraries(${NVML_CLIENT_OUTPUT} PRIVATE stdc++)
-        target_link_options(${NVML_CLIENT_OUTPUT} PRIVATE
-            "LINKER:--version-script=${CMAKE_CURRENT_SOURCE_DIR}/nvml.exports"
-        )
-      endif()
+    add_library(${NVML_CLIENT_OUTPUT} SHARED ${NVML_CLIENT_SOURCES})
+    target_include_directories(${NVML_CLIENT_OUTPUT} PRIVATE ${CUDAToolkit_INCLUDE_DIRS})
+    if(OpenSSL_FOUND)
+      target_compile_definitions(${NVML_CLIENT_OUTPUT} PRIVATE LUPINE_TLS_OPENSSL)
+      target_link_libraries(${NVML_CLIENT_OUTPUT} PRIVATE OpenSSL::SSL OpenSSL::Crypto)
+    endif()
+    target_link_libraries(${NVML_CLIENT_OUTPUT} PRIVATE ${NGHTTP2_LIBRARY} lupine_lz4)
+    if(NOT APPLE)
+      target_link_libraries(${NVML_CLIENT_OUTPUT} PRIVATE stdc++)
+      target_link_options(${NVML_CLIENT_OUTPUT} PRIVATE
+          "LINKER:--version-script=${CMAKE_CURRENT_SOURCE_DIR}/nvml.exports"
+      )
+    endif()
+    set_target_properties(${NVML_CLIENT_OUTPUT} PROPERTIES
+        POSITION_INDEPENDENT_CODE ON
+        OUTPUT_NAME nvidia-ml
+    )
+    if(NOT APPLE)
       set_target_properties(${NVML_CLIENT_OUTPUT} PROPERTIES
-          POSITION_INDEPENDENT_CODE ON
-          OUTPUT_NAME nvidia-ml
           VERSION 1
           SOVERSION 1
       )
     endif()
 endif()
 
-if(LUPINE_BUILD_SERVER)
+if(NOT APPLE)
 add_executable(${SERVER_OUTPUT} ${SERVER_SOURCES} ${SERVER_HEADERS})
 target_include_directories(${SERVER_OUTPUT} PRIVATE ${CUDAToolkit_INCLUDE_DIRS})
 target_compile_definitions(${SERVER_OUTPUT} PRIVATE
@@ -363,7 +360,7 @@ else()
 endif()
 endif()
 
-if (LUPINE_BUILD_SERVER AND NOT WIN32)
+if (NOT WIN32 AND NOT APPLE)
     target_compile_options(${SERVER_OUTPUT} PRIVATE
         $<$<COMPILE_LANGUAGE:CUDA>: --compiler-options=-fPIC,-g,-Wno-deprecated-declarations>
     )
@@ -383,6 +380,6 @@ if(TARGET ${NVML_CLIENT_OUTPUT})
     message(STATUS "Building NVML client output: ${NVML_CLIENT_OUTPUT}")
 endif()
 
-if(LUPINE_BUILD_SERVER)
+if(NOT APPLE)
   message(STATUS "Building server output: ${SERVER_OUTPUT}")
 endif()
