+++ cuda_client.cpp
@@ -55,9 +55,9 @@
 #include "events.h"
 #include "ipc.h"
 #include "lupine_attr_sizes.h"
-#include "lupine_elf.h"
 #include "lupine_fatbin.h"
 #include "lupine_log.h"
+#include "lupine_platform.h"
 #include "memcpy.h"
 #include "rpc.h"
 #include "third_party/libcuckoo/libcuckoo/cuckoohash_map.hh"
